修复多处逻辑bug和兼容性问题
ml/:
- labels: triple barrier超时退出使用真实1min收盘价而非ffill聚合价
- labels: 同bar双触碰止盈止损时用收盘价判断方向,消除系统性偏差
- train: forward_bars回退键与build_labels保持一致(兼容forward_days)
- evaluate: Sharpe计算forward_bars=1与>1执行延迟统一(去掉多余shift)
- evaluate: FU每日交易分钟数从240修正为345(夜盘120+早盘75+午前60+午后90)
- predict: NaN过滤从dropna(how=all)改为>50%非空,与训练一致
- train/predict/backtest: yaml.unsafe_load改为safe_load

futuresquant/:
- account: mark_to_market加回冻结保证金,修复持仓期间权益虚低
- engine: 回测结束平仓后权益点时间戳加1ms避免重复
- storage: volume/open_interest从int32改为int64防溢出
- momentum/volume: 移除pandas 2.2+废弃的fill_method参数
- futures_structure: 合约年份分界从<=30改为<70,支持到2069年

research/:
- 02: 补充缺失的calc_rolling_ic函数
- 02/06: 净值计算shift逻辑与ml/evaluate.py统一

+++ futuresquant/data/storage.py
@@ -142,7 +142,7 @@ class ParquetCache:
                 df[col] = df[col].astype("float32")
         for col in ("volume", "open_interest"):
             if col in df.columns:
-                df[col] = df[col].astype("int32")
+                df[col] = df[col].astype("int64")
 
         df.to_parquet(parquet_path, engine="pyarrow", compression="snappy")
         logger.debug("Cached %d rows → %s", len(df), parquet_path)
